add auto play

+++ app/src/main/java/com/example/freshcard/Structure/ResultTest.kt
@@ -2,4 +2,4 @@ package com.example.freshcard.Structure
 
 import com.google.type.DateTime
 
-data class ResultTest(val idUser: String, val amountCorrect: Int, val duration: Int, val time: DateTime, val type: String){}
+data class ResultTest(val idUser: String, val amountCorrect: Int, val duration: Long, val time: DateTime, val type: String){}
